+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.0.9
+- Add `/latest/<game name>` URL to allow refreshing the viewer to reload the latest save.
+- Update interface to list latest download and view URLs.
+
 ## 0.0.8
 - Add version printing with `-v` or `--version`.